add[new feature] => invoice pdf feature

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const {validationResult} = require('express-validator');
 
 const Product = require('../models/product');
+const FileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -154,6 +155,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (image) {
+                FileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             return product.save();
@@ -192,10 +194,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({
-        _id: prodId,
-        userId: req.user._id
-    })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            FileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({
+                _id: prodId,
+                userId: req.user._id
+            });
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
